fix(catalog): fix catalog
ref: MANAGER-11852
