+++ packages/manager/apps/catalog/README.md
@@ -1,3 +1,3 @@
 # @ovh-ux/manager-catalog-app
 
-> catalog revamp react app
+> OVHcloud Catalog App
