--- packages/manager/apps/catalog/src/catalog-revamp.config.ts
@@ -1,1 +1,0 @@
-export default {};
