+++ packages/manager/apps/catalog/src/hooks/useCatalog.ts
@@ -3,35 +3,9 @@ import { useQuery, UseQueryResult } from '@tanstack/react-query';
 import {
   getManagerHubCatalogList,
   getManagerHubCatalogListQueryKey,
+  Product,
 } from '@/api';
 import { filterProducts } from '@/utils/utils';
-
-export type Product = {
-  id: number;
-  name: string;
-  description: string;
-  lang: string;
-  categories: string[];
-  url: string;
-  regionTags: string[];
-  productName: string;
-  order: string;
-  universe: string;
-  category: string;
-};
-
-export type CatalogData = {
-  data: {
-    catalog: {
-      data: Product[];
-    };
-  };
-};
-
-interface ServiceData {
-  status: number;
-  data: CatalogData;
-}
 
 interface UseCatalogParams {
   categories?: string[];
@@ -47,7 +21,7 @@ export const useCatalog = ({
   const [products, setProducts] = useState<Product[]>([]); // full list of products
   const [results, setResults] = useState<Product[]>([]); // the filtered list of products
 
-  const service: UseQueryResult<ServiceData> = useQuery({
+  const service = useQuery({
     queryKey: getManagerHubCatalogListQueryKey,
     queryFn: () => getManagerHubCatalogList(),
     staleTime: Infinity,
@@ -56,8 +30,8 @@ export const useCatalog = ({
   const { error } = service;
 
   useEffect(() => {
-    if (service.isSuccess && service.data.status === 200) {
-      const response: Product[] = service.data?.data?.data?.catalog?.data || [];
+    if (service?.data?.status === 200) {
+      const response: Product[] = service.data?.data;
       setProducts(response);
       setResults(response);
     }
